Notes: Refreshing the browser inside the cart with orders inside are still causing the error and not being able to render. Also, there are problem with communcation with back-end items. There are some Uncaught items that are still being looked at.
Prop-drilling from 'Order Now' component to all children components
could be the cause of the error. So, it will be look at and props will
be replaced using context so they can be little bit more organized.

Cart.js have to modified so that users can submit orders. Also, that
it is modified so that the app waits till the network request to the
back-end resources to complete before it completes the order and
preventing it from making any inadvertent submissions and creating any
unwanted orders from network connection.

Alert.js have been added to let users know that the order is complete or
failed due to some errors. However, since the manipulation of the
back-end resource from the very start of the project, I may have to look
at APIs that arr connected to orders.

Employee of the month discount feature needs to be modified so that
it is showing the real life situations.

Some of 'Order Now' component's child components especailly that are
playing main roles in 'Pack-orders' are unsaved due to the
error fixing. They have been causing more confusions. So it will be
looked at again later. And some features will be removed.

Changes to be committed:
	new file:   src/components/Alert.css
	new file:   src/components/Alert.js
	modified:   src/components/Cart.js
	modified:   src/reducers/cartReducer.js