Fix the critical error that caused the application to crash when the browser is ordered to refresh/reload. It was due to the App.js completing its tasks before it fetched all neccesary data from the backend resource. Added some features so that users can login. Some users like 'associate' can look at special components however, it is not yet to be done.
Future Implements:
Need to make the special orders to work in the Order Now. Currently,
'Pack Order' and 'Gift Order' is crashing.
I still need to lift props and convert prop drilling to context to
make the application easy to maintain. This will be the biggest update
Also, need to implment sidebars, account settings.
Displaying orders for applcation manager will be needed.

Changes to be committed:
	modified:   src/App.js
	modified:   src/components/Cart.js
	modified:   src/components/Header.js
	new file:   src/components/Login.css
	new file:   src/components/Login.js
	new file:   src/components/Orders.css
	new file:   src/components/Orders.js
	new file:   src/components/UserDetails.css
	new file:   src/components/UserDetails.js
	new file:   src/context/CurrentUserContext.js